// src/app/core/enums/dropdown-options.enum.ts
export enum Ethnicity {
    IP_HOUSEHOLD = 1,
    NON_IP_HOUSEHOLD = 2
}
  
export enum SocioeconomicStatus {
    NHTS_4PS = 1,
    NHTS_NON_4PS = 2,
    NON_NHTS = 3
}
  
export enum TypeOfWaterSource {
    LEVEL_I = 1,
    LEVEL_II = 2,
    LEVEL_III = 3,
    OTHERS = 4
}
  
export enum TypeOfToiletFacility {
    A = 1,
    B = 2,
    C = 3,
    D = 4,
    E = 5,
    F = 6,
    G = 7
}
  
export enum RelationshipToHeadOfHousehold {
    HEAD = 1,
    SPOUSE = 2,
    SON = 3,
    DAUGHTER = 4,
    OTHERS = 5
}
  
export enum Gender {
    MALE = 1,
    FEMALE = 2
}
  
export enum CivilStatus {
    SINGLE = 1,
    MARRIED = 2,
    WIDOW = 3,
    SEPARATED = 4,
    COHABITATION = 5
}
  
export enum MemberType {
    MEMBER = 1,
    DEPENDENT = 2
}
  
export enum PhilhealthCategory {
    FEP = 1,
    FEG = 2,
    IE = 3,
    NHTS = 4,
    SC = 5,
    IP = 6,
    UNKNOWN = 7
}
  